% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddConfInterval.R
\name{AddConfInterval}
\alias{AddConfInterval}
\title{Add confidence intervals to a spectrum}
\usage{
AddConfInterval(spec, MINVALUE = 1e-10, pval = 0.05)
}
\arguments{
\item{spec}{spectrum list(spec,freq,dof)}

\item{MINVALUE}{Minimum value to which the confidence interval is limited}

\item{pval}{Interval from (pval/2 to 1-pval/2) is constructed}
}
\value{
spectrum as the input but including lim.1 and lim.2 as new list elements
}
\description{
Add confidence intervals to a spectrum
}
\author{
Thomas Laepple
}

