% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPowerlaw.R
\name{AnPowerlaw}
\alias{AnPowerlaw}
\title{A PSD(freq) for a powerlaw with variance 1}
\usage{
AnPowerlaw(beta, freq, return.scaling = FALSE)
}
\arguments{
\item{beta}{slope of the powerlaw}

\item{freq}{frequency vector}
}
\value{
vector containing the PSD
}
\author{
Thomas Laepple
}

