% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxNearest.R
\name{ApproxNearest}
\alias{ApproxNearest}
\title{approximate a timeseries using the nearest neighbour}
\usage{
ApproxNearest(x, y, xout, rule = 1)
}
\arguments{
\item{x}{numeric vector giving the coordinates of the points to be interpolate}

\item{y}{corresponding y values}

\item{xout}{set of numeric values specifying where interpolation is to take place.}

\item{rule}{an integer (of length 1 or 2) describing how interpolation is
to take place outside the interval see ?approx}
}
\value{
a list with components 'x' and 'y', containing
length(xouth) coordinates which interpolate the given data points
}
\description{
extends approx which always takes the right or left neighbour 
or the weighted mean between both if f>0<1
}
\examples{
x<-1:10
y<-1:10
xout<-seq(from=0,to=11,by=0.01)
plot(x,y,type="b",pch=19,xlim=range(xout))
result<-ApproxNearest(x,y,xout)
lines(result,col="red")
}
\author{
Thomas Laepple
}
