% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpec.R
\name{as.data.frame.spec}
\alias{as.data.frame.spec}
\title{Transform a spec Object Into a Dataframe}
\usage{
\method{as.data.frame}{spec}(x)
}
\arguments{
\item{x}{A spec object}
}
\value{
A dataframe or tibble (if package tibble is installed)
}
\description{
Transform a spec Object Into a Dataframe
}
\examples{
library(PaleoSpec)
ts1 <- ts(rnorm(100))
sp1 <- SpecMTM(ts1)
sp1_df <- as.data.frame(sp1)
}
\seealso{
Other TidySpec: 
\code{\link{DF2Spec}()},
\code{\link{Spec2DF}()},
\code{\link{as.spec}()},
\code{\link{as_spec_df}()}
}
\concept{TidySpec}
