% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpec.R
\name{as.spec}
\alias{as.spec}
\title{Make a spec Object}
\usage{
as.spec(x)
}
\arguments{
\item{x}{A named list with (at least) the elements freq, spec, and dof}
}
\value{
Object of class spec
}
\description{
Make a spec Object
}
\seealso{
Other TidySpec: 
\code{\link{DF2Spec}()},
\code{\link{Spec2DF}()},
\code{\link{as.data.frame.spec}()},
\code{\link{as_spec_df}()}
}
\concept{TidySpec}
