% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpec.R
\name{as_spec_df}
\alias{as_spec_df}
\title{Make a spec_df Object}
\usage{
as_spec_df(x)
}
\arguments{
\item{x}{A dataframe or tibble with (at least) columns freq, spec, and dof}
}
\value{
Object of class spec_df
}
\description{
Make a spec_df Object
}
\seealso{
Other TidySpec: 
\code{\link{DF2Spec}()},
\code{\link{Spec2DF}()},
\code{\link{as.data.frame.spec}()},
\code{\link{as.spec}()}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
\concept{TidySpec}
