% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvgToBin.R
\name{AvgToBin}
\alias{AvgToBin}
\title{Bin averaging}
\usage{
AvgToBin(x, y, N = 2, breaks = pretty(x, N), bFill = FALSE)
}
\arguments{
\item{x}{vector of values on which the data in \code{y} is tabulated;
e.g. depth or time points.}

\item{y}{vector of observation values to be averaged into bins. Must have the
same length as \code{x}.}

\item{N}{desired number of breaks (ignored if \code{breaks} are supplied
directly).}

\item{breaks}{vector of break point positions to define the averagig bins; if
omitted, break point positions are calculated from the range of \code{x}
and the desired number of breaks given by \code{N}.}

\item{bFill}{logical; if \code{TRUE}, fill empty bins using linear
interpolation from the neighbours to the center of the bin.}
}
\value{
a list with four elements:
\describe{
\item{\code{breaks}:}{numeric vector of the used break point positions.}
\item{\code{centers}:}{numeric vector with the positions of the bin centers.}
\item{\code{avg}:}{numeric vector with the bin-averaged values.}
\item{\code{nobs}:}{numeric vector with the number of observations
  contributing to each bin average.}
}
}
\description{
Average a vector into bins.
}
\details{
This function averages the vector \code{y} into bins according to the positon
of \code{x} within the breaks. You can either specify a desired number N of
breaks which are used to calculate the actual breaks via \code{pretty(x, N)},
or directly specify the N + 1 break positions. The averaging bins are defined
via \code{x > breaks[i]} and \code{x <= breaks[i + 1]}. If \code{bFill =
TRUE}, empty bins are filled using linear interpolation from the neighbours
to the center of the bin.

Probably the binning could be considerably speeded up by using \code{?cut}.
}
\author{
Thomas Laepple
}
