% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecACF.R
\name{BinTimeseries}
\alias{BinTimeseries}
\title{Bin a Timeseries Preserving Empty Bins}
\usage{
BinTimeseries(
  time.var,
  value.var,
  bin.width,
  strt.time = NULL,
  end.time = NULL
)
}
\arguments{
\item{time.var}{time variable, a vector}

\item{value.var}{value variable, a vector}

\item{bin.width}{with of time bins in output}

\item{strt.time}{first time value in output, defaults to min(time.var) - bin.width/2}

\item{end.time}{last time values in output, defaults to max(time.var) + bin.width/2}
}
\value{
a dataframe
}
\description{
Convert a continuous time series to a discrete (regular)
  timeseries by binning, preserving any empty bins. Additionally returns the
  number of data points in each bin
}
\examples{
set.seed(20230312)
x <- cumsum(rgamma(20, shape = 1.5, rate = 1.5/10))
y <- SimProxySeries(a = 0.1, b = 1, t.smpl = x)
BinTimeseries(x, y, bin.width = 15)
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
