% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAvgSpectra.R
\name{CalcAvgSpectra}
\alias{CalcAvgSpectra}
\title{average spectra with weighting}
\usage{
CalcAvgSpectra(specList, iRemoveLowest = 1, weights = rep(1,
  length(specList)))
}
\arguments{
\item{iRemoveLowest}{number of lowest frequencies to remove (e.g. to remove detrending bias)}

\item{weights}{vector of weights (same length as elements in speclist)}

\item{speclist}{list of spectra}
}
\value{
list(spec,nRecords) spec=average spectrum, nRecords = number of records contributing to each spectral estimate
}
\description{
average spectra with weighting, spectra can have different resolution and span a different freq range
}
\details{
Calculate the weighted mean spectrum of all spectra by interpolating them to the highest resolution
}
\author{
Thomas Laepple
}

