% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coherence.R
\name{Coherence}
\alias{Coherence}
\title{MTM coherence estimator}
\usage{
Coherence(
  x,
  y = NULL,
  dt = 1,
  nw = 8,
  k = NULL,
  detrend = T,
  qbias = 0,
  confn = 0
)
}
\arguments{
\item{x, y}{input data vectors (or array if y=NULL)}

\item{dt}{time stepping (default = 1)}

\item{nw}{number of Slepian taper (default = 8)}

\item{k}{order (default = NULL)}

\item{detrend}{(default=T)}

\item{qbias}{bias correction fo coherence estimate (default = 0/no)}

\item{confn}{number of interactions to use in MC for uncertainty estimation
of the phase (default = 0)}

\item{dpssIN}{allows to use an predefined dpss object to reduce computation
time when applied multiple times (default=NULL). If specified nw and k are
obsolete.}
}
\value{
spectra object
}
\description{
estimates coherence using slepian tapers with adaptive
  weighting. Function is an adaption of Peter Huybers Matlab coherence
  function cmtm.m
}
