% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColTransparent.R
\name{ColTransparent}
\alias{ColTransparent}
\title{Modify a color to get brighter and tranparent for the confidence intervals}
\usage{
ColTransparent(color, alpha = 0.8, beta = 150)
}
\arguments{
\item{color}{color value, e.g. "red"}

\item{alpha}{(0..1) transparency value}

\item{beta}{(0..255) to make it brighter, this value gets added on the RGB values}
}
\value{
modified color
}
\description{
Modify a color to get brighter and tranparent for the confidence intervals
}
\author{
Thomas Laepple
}

