% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfRatio.R
\name{ConfRatio}
\alias{ConfRatio}
\title{Confidence Interval of ratios}
\usage{
ConfRatio(varratio, df.1, df.2, pval = 0.1)
}
\arguments{
\item{varratio}{}

\item{df.1}{degree of freedom of denominator}

\item{df.2}{degree of freedom of numerator}

\item{pval}{}
}
\value{
lower and upper confidence intervals
}
\description{
Confidence Interval of ratios
based on a ChiSquare Distribution
}
\author{
Thomas Laepple
}
