% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfVar.R
\name{ConfVar}
\alias{ConfVar}
\title{Provide ChiSquared confidence intervals for ratios}
\usage{
ConfVar(varlist, pval = 0.05)
}
\arguments{
\item{varlist}{list(var,dof)}

\item{pval}{requested p-value}
}
\value{
Output: confidence intervals
}
\description{
Provide ChiSquared confidence intervals for ratios
}
\author{
Thomas Laepple
}

