% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpec.R
\name{DF2Spec}
\alias{DF2Spec}
\title{Transform a spec_df Object into a spec Object or List of spec Objects}
\usage{
DF2Spec(spec_df)
}
\arguments{
\item{spec_df}{A spec_df object}
}
\value{
A spec object or a list of spec objects
}
\description{
Transform a spec_df Object into a spec Object or List of spec Objects
}
\examples{
library(PaleoSpec)
ts1 <- ts(SimPLS(1000, beta = 1))
sp1 <- SpecMTM(ts1)
ts2 <- ts(rnorm(1000))
sp2 <- SpecMTM(ts2)
sp_lst <- list(sp1 = sp1, sp2 = sp2)
sp_df <- Spec2DF(sp_lst)
sp_df
str(DF2Spec(sp_df))
}
\seealso{
Other TidySpec: 
\code{\link{Spec2DF}()},
\code{\link{as.data.frame.spec}()},
\code{\link{as.spec}()},
\code{\link{as_spec_df}()}
}
\author{
Andrew Dolman <andrew.dolman@awi.de>
}
\concept{TidySpec}
