% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimProxySeries.R
\name{gen.ran.seq}
\alias{gen.ran.seq}
\title{Generate Random Numbers Without Affecting the State of .Random.seed}
\usage{
gen.ran.seq(n, nser, s)
}
\arguments{
\item{n}{number of observations}

\item{nser}{number of replicate series}

\item{s}{NA or integer, passed to set.seed() if integer}
}
\value{
a matrix
}
\description{
Generate Random Numbers Without Affecting the State of .Random.seed
}
\keyword{internal}
