% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetVarFromSpectra.R
\name{GetVarFromSpectra}
\alias{GetVarFromSpectra}
\title{Variance estimate by integrating a part of the spectrum}
\usage{
GetVarFromSpectra(spec, f, dfreq = 0.05, df.log = 0, bw = 3)
}
\arguments{
\item{spec}{spectrum (list of spec,freq,dof) to be analysed}

\item{f}{f[1],f[2]: frequency interval to be analysed}

\item{dfreq}{frequency discretisation used in the temporary interpolation}

\item{df.log}{if > 0, smooth the spectra prior to integrating}

\item{bw}{the bandwidth assumed for the confinterval calculation (from the multitaper spectral estimate)}
}
\value{
list(var,dof)  variance and corresponding dof
}
\description{
Variance estimate by integrating a part of the spectrum
}
\author{
Thomas Laepple
}

