% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InverseFilter.R
\name{InverseFilter}
\alias{InverseFilter}
\title{Construct the inverse filter in the time domain}
\usage{
InverseFilter(filter.weights)
}
\arguments{
\item{filter.weights}{}
}
\value{
filter weights for the inverse filter
}
\description{
Construct the inverse filter in the time domain
}
\author{
Thomas Laepple
}
