% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is.spectrum}
\alias{is.spectrum}
\title{Check for spectral object}
\usage{
is.spectrum(x, check.only = FALSE, dof = TRUE)
}
\arguments{
\item{x}{an object to be checked.}

\item{check.only}{logical; the default means to issue an error when a
requirement is not met. If set to \code{TRUE}, all requirements are
checked, and the function returns \code{TRUE} if they all pass, or
\code{FALSE} if one or more checks fail.}

\item{dof}{logical; per default, \code{x} is also checked for containing a
\code{dof} vector, which can be turned off here when unnecessary.}
}
\value{
a logical value for \code{check.only = TRUE}.
}
\description{
In \code{PaleoSpec} a "spectral object" is defined to be a list consisting as
the minimum requirement of the three elements \code{freq}, \code{spec}, and
\code{dof}, which are numeric vectors of the same length giving the frequency
axis, the power spectral density, and the degrees of freedom of the spectral
estimate. Most functions in this package require such an object as their main
input. This function checks an object for the minimum requirements.
}
\author{
Thomas Münch
}
