% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastElement.R
\name{LastElement}
\alias{LastElement}
\title{last element of a vector}
\usage{
LastElement(x)
}
\arguments{
\item{x}{}
}
\value{
last element of X
}
\description{
last element of a vector
}
\author{
Thomas Laepple
}
