% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogSmooth.R
\name{LogSmooth}
\alias{LogSmooth}
\title{Smoothes the spectrum using a log smoother}
\usage{
LogSmooth(spectra, df.log = 0.05, removeFirst = 1e+06, removeLast = 0,
  bLog = FALSE)
}
\arguments{
\item{spectra}{spectra: list(spec,freq)
spec[specIndex]: spectra density vector
freq[specIndex]: frequency vector}

\item{df.log}{width of the smoother in log units}

\item{removeFirst}{elements to remove on the slow side (one element recommended because of the detrending}

\item{removeLast}{elements to remove on the fast side}

\item{bLog}{TRUE: average in the log space of the power, FALSE: arithmetic average}
}
\value{
smoothed spectrum
}
\description{
Smoothes the spectrum using a log smoother.
}
\author{
Thomas Laepple
}

