% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPlot.R
\name{LPlot}
\alias{LPlot}
\title{add Logplot + transparent confidence interval for the spectral plotting}
\usage{
LPlot(x, conf = TRUE, bPeriod = FALSE, col = "black", alpha = 0.3,
  removeFirst = 0, removeLast = 0, xlab = "f", ylab = "PSD",
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{spectra object}

\item{conf}{TRUE: Plot confidence interval}

\item{bPeriod}{if \code{TRUE} the x-axis is displayed in units of period
(inverse frequency), increasing to the left. Defaults to \code{FALSE}.}

\item{col}{color}

\item{alpha}{transparency}

\item{removeFirst}{omit removeFirst values on the low frequency side}

\item{removeLast}{omit removeFirst values on the high frequency side}

\item{xlab}{label of x-axes}

\item{ylab}{label of y-axes}

\item{xlim}{range of x-axis values; if \code{NULL} (the default) it is
calculated internally and automatically reversed for \code{bPeriod =
TRUE}.}

\item{ylim}{range of y-axis values; if \code{NULL} (the default) it is
calculated internally.}

\item{...}{other parameters to be passed to the line functio}
}
\value{
none
}
\description{
add Logplot + transparent confidence interval for the spectral plotting
}
\examples{
x<-ts(arima.sim(list(ar = 0.9),1000))
spec<-SpecMTM(x)
LPlot(spec,col="grey")
LLines(LogSmooth(spec),lwd=2)
}
\author{
Thomas Laepple
}
