% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEquidistant.R
\name{MakeEquidistant}
\alias{MakeEquidistant}
\title{Average an irregular timeseries to a regular timeseries}
\usage{
MakeEquidistant(t.x, t.y, dt = 0.1, time.target = seq(from = t.x[1], to =
  t.x[length(t.x)], by = dt), dt.hres = NULL, bFilter = TRUE, k = 5,
  kf = 1.2)
}
\arguments{
\item{t.x}{vector of timepoints}

\item{t.y}{vector of corresponding values}

\item{dt}{target timestep; can be omitted if time.target is supplied}

\item{time.target}{time vector to which timeseries should be averaged/interpolated to
by default the same range as t.x with a timestep dt}

\item{dt.hres}{timestep of the intermediate high-resolution interpolation.
Should be smaller than the smallest timestep}

\item{bFilter}{(TRUE) low passs filter the data to avoid aliasing, (FALSE) just interpolate}

\item{k}{scaling factor for the Length of the filter (increasing creates}

\item{kf}{scaling factor for the lowpass frequency; 1 = Nyquist, 1.2 =
1.2xNyquist is a tradeoff between reducing variance loss and keeping
aliasing small}
}
\value{
ts object with the equidistant timeseries
}
\description{
Average an irregular timeseries to a regular timeseries
}
\details{
Make an irregular timeseries equidistant by
interpolating to high resolution, lowpass filtering to the Nyquist
frequency, and subsampling; e.g. as used in Huybers and Laepple, EPSL 2014
}
\author{
Thomas Laepple
}

