% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecACF.R
\name{mean.ACF}
\alias{mean.ACF}
\title{Title}
\usage{
\method{mean}{ACF}(x, return.working = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{return.working}{return intermediate steps, defaults to FALSE}
}
\value{
a matrix
}
\description{
Title
}
\keyword{internal}
