% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSpectrum.R
\name{MeanSpectrum}
\alias{MeanSpectrum}
\title{average spectra, optionally with weighting and interpolation}
\usage{
MeanSpectrum(specList, iRemoveLowest = 0, weights = rep(1,
  length(specList)))
}
\arguments{
\item{iRemoveLowest}{number of lowest frequencies to remove (e.g. to remove detrending bias)}

\item{weights}{vector of weights (same length as elements in speclist)}

\item{speclist}{list of spectra}
}
\value{
list(spec,nRecords) spec=average spectrum, nRecords = number of records contributing to each spectral estimate
}
\description{
average spectra with weighting, spectra can have different resolution and span a different freq range, in this case, they get linearly interpolate to the same (maximum range and highes resolution) frequency resolution first.
}
\examples{
spec1<-SpecMTM(ts(rnorm(100),deltat=1))
spec2<-SpecMTM(ts(rnorm(100),deltat=1))
spec3<-SpecMTM(ts(rnorm(100),deltat=2))
spec.avg.12<-MeanSpectrum(list(spec1,spec2))
spec.avg.123<--MeanSpectrum(list(spec1,spec2,spec3)) 
}
\author{
Thomas Laepple
}

