% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsampleTimeseriesBlock.R
\name{MonthlyFromDaily}
\alias{MonthlyFromDaily}
\title{Bin daily values to monthly values}
\usage{
MonthlyFromDaily(ts.daily)
}
\arguments{
\item{ts.daily}{vector of 365 values}
}
\value{
vector of 12 values
}
\description{
Assumes months of equal length and a 365 day long year
}
\author{
Thomas Laepple
}
