% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSP.CorAfterRollmean.R
\name{PSP.CorAfterRollmean}
\alias{PSP.CorAfterRollmean}
\title{Numerical correlation of random timeseries with different filtering (running mean)}
\usage{
PSP.CorAfterRollmean(N, betaSignal, betaNoise, R)
}
\arguments{
\item{N}{Number of points of the timeseries}

\item{betaSignal}{powerlaw slope of the signal}

\item{betaNoise}{powerlaw slope of the noise}

\item{R}{expected correlation of the whole timeseries}
}
\value{
correlation of unfiltered, 10point mean and 50 point mean
}
\description{
Numerical correlation of random timeseries with different filtering (running mean)
}
\examples{
temp <- replicate(1000,PSP.CorAfterRollmean(1000,1,0,0.5))
rowMeans(temp)
PSP.CorUntilF(c(0.5,0.5/10,0.5/50),1,0,1000,0.5)
}
\author{
Thomas Laepple
}
