% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSP.CorUntilF.R
\name{PSP.CorUntilF}
\alias{PSP.CorUntilF}
\title{lowpass filtered expected correlation of powerlaw signal pair}
\usage{
PSP.CorUntilF(f, betaSignal, betaNoise, N, r)
}
\arguments{
\item{f}{frequency until which to integrate}

\item{betaSignal}{powerlaw slope of the signal}

\item{betaNoise}{powerlaw slope of the noise}

\item{N}{Number of points per timeseries}

\item{r}{expected correlation of the unfiltered}
}
\value{
expected correlation of the timeseries
}
\description{
Correlation of two timeseries with powerlaw signal and powerlaw
noise evaluated until f
this equals the correlation of linearly coupled lowpass filtered powerlaw
process
}
\examples{
temp <- replicate(1000,PSP.CorAfterRollmean(1000,1,0,0.5))
rowMeans(temp)
PSP.CorUntilF(f=c(0.5,0.5/10,0.5/50),betaSignal=1,betaNoise=0,N=1000,r=0.5)
}
\author{
Thomas Laepple
}
