% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSpectrum.R
\name{remove.lowestFreq}
\alias{remove.lowestFreq}
\title{Remove low frequencies}
\usage{
remove.lowestFreq(spec, iRemove)
}
\arguments{
\item{spec}{a list object of class \code{"spec"}.}

\item{iRemove}{integer; number of lowest frequencies to remove.}
}
\value{
the input data with the data at the \code{iRemove} lowest
  frequencies removed.
}
\description{
Remove the lowest frequencies from a spectra, since they are biased by
detrending and the MTM estimator.
}
\author{
Thomas Laepple
}
