% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPowerlaw.R
\name{SimFromEmpiricalSpec}
\alias{SimFromEmpiricalSpec}
\title{Simulate a random timeseries consistent with an arbitrary numerical power spectrum}
\usage{
SimFromEmpiricalSpec(spec, N)
}
\arguments{
\item{spec}{Numerical power spectrum consisting of a list with components $freq and $spec}

\item{N}{length of timeseries to be generated}
}
\value{
vector containing the timeseries
}
\description{
Adapted from SimPowerlaw
}
\examples{
# Create a piecewise spectrum

## helper function to generate continuous piecewise spectrum
new.scl <- function(f, prev.scl, new.slp, prev.slp){
  prev.scl*f^(prev.slp-new.slp)
}

emp.spec <- data.frame(freq = seq(1/1e05, 1/2, 1/1e05))
emp.spec$spec <- with(emp.spec, {
  0.1 * freq ^ -1 * (freq > 1e-02) +
    new.scl(1e-02, 0.1,-2,-1) * freq ^ -2 * (freq <= 1e-02 & freq > 1e-03) +
    new.scl(1e-03, new.scl(1e-02, 0.1,-2,-1),-1,-2) * freq ^ -1 * (freq <= 1e-03)

})

plot(emp.spec, type = "l", log = "xy")

# Sample consistent with spectrum
ts1 <- ts(SimFromEmpiricalSpec(emp.spec, 50000))

# re-estimate power spectrum
spec1 <- SpecMTM(ts1)
LPlot(spec1)
lines(emp.spec, col = "Red")
abline(v = c(1e-02, 1e-03), col = "Green")
}
\author{
Thomas Laepple and Andrew Dolman
}
