% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPowerlawPiecewise.R
\name{SimPowerlawPiecewise}
\alias{SimPowerlawPiecewise}
\title{#Simulate a timeseries with length N which has a spectra consisting of two powerlaws}
\usage{
SimPowerlawPiecewise(beta1, beta2, N, deltat = 1, breakpoint = 1/50)
}
\arguments{
\item{beta1}{slope for frequencies lower than breakpoint}

\item{beta2}{slope for frequencies higher than breakpoint}

\item{N}{Number of points to simulate}

\item{deltat}{timestep of the timeseries}

\item{breakpoint}{frequency of the breakpoint}
}
\value{
N random numbers drawn according to the piecewise powerlaw PSD
}
\author{
Thomas Laepple
}
