% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogSmooth.R
\name{smoothlin.cutEnd}
\alias{smoothlin.cutEnd}
\title{smoothlin.cutEnd}
\usage{
smoothlin.cutEnd(x, f, df.log, dof = 1)
}
\arguments{
\item{x}{}

\item{f}{}

\item{df.log}{}

\item{dof}{}
}
\value{
smoothed x
}
\author{
Thomas Laepple
}
