% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogSmooth.R
\name{smoothlog.cutEnd}
\alias{smoothlog.cutEnd}
\title{smoothlog.cutEnd}
\usage{
smoothlog.cutEnd(x, f, df.log, dof = 1)
}
\value{
smoothed x
}
\description{
smoothlog.cutEnd
}
\author{
Thomas Laepple
}
\keyword{internal}
