% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidySpec.R
\name{Spec2DF}
\alias{Spec2DF}
\title{Transform Spec Object(s) Into a Dataframe}
\usage{
Spec2DF(x)
}
\arguments{
\item{x}{A spec object or (optionally named) list of spec objects.}
}
\value{
A spec_df object
}
\description{
Transform Spec Object(s) Into a Dataframe
}
\examples{
library(PaleoSpec)
ts1 <- ts(SimPLS(1000, beta = 1))
sp1 <- SpecMTM(ts1)
sp1 <- AddConfInterval(sp1)
ts2 <- ts(rnorm(1000))
sp2 <- SpecMTM(ts2)
sp_lst <- list(sp1 = sp1, sp2 = sp2)
sp_df <- Spec2DF(sp_lst)
sp_df
}
\seealso{
Other TidySpec: 
\code{\link{DF2Spec}()},
\code{\link{as.data.frame.spec}()},
\code{\link{as.spec}()},
\code{\link{as_spec_df}()}
}
\concept{TidySpec}
