% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecMean.R
\name{SpecMean}
\alias{SpecMean}
\title{Mean spectrum from of a list of spectra}
\usage{
SpecMean(speclist, weight = FALSE, dof = TRUE)
}
\arguments{
\item{speclist}{list of spectra each containing list(spec, freq, dof)}

\item{weight}{weight by the uncertainty}

\item{dof}{degrees of freedom}
}
\value{
the mean spectra
}
\description{
Mean spectrum from of a list of spectra
}
\details{
Returns the mean of the spectra
Inputs:
speclist[[]], each containing
a list(spec,freq,dof)
                    spec[specIndex]: spectra density vector
                    freq[specIndex]: frequency vector
                    dof[specIndex]: DOF ... or a single value df
weight:             weight by the uncertainty

Weighting by 1/sigma^2;  sigma^2 = variance  is proportional to 1/DOF

This is true if we assume a constant spectral density... if not,
higher spectral densities have higher uncertainty
}
