% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimProxySeries.R
\name{Torb2Spec}
\alias{Torb2Spec}
\title{Convert Torben Kunz' spec output to standard PaleoSpec spec object}
\usage{
Torb2Spec(x)
}
\arguments{
\item{x}{}
}
\value{
spec object
}
\description{
Convert Torben Kunz' spec output to standard PaleoSpec spec object
}
\keyword{internal}
