% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecACF.R
\name{TrimNA}
\alias{TrimNA}
\title{Remove leading and trailing rows of all NA}
\usage{
TrimNA(m, trim = c("all", "any"))
}
\arguments{
\item{m}{a numeric matrix, data.frame or vector}

\item{trim}{trim leading and trailing rows of "all" NA or containing "any" NA values}
}
\value{
a numeric matrix, data.frame or vector
}
\description{
Remove leading and trailing rows of all NA
}
\examples{
m <- matrix(c(NA, NA, NA, 1, NA, NA, NA, 1, 1, NA, NA, NA, 1:9, NA,NA,NA, 10:12, NA, 1, NA, NA,NA,NA), ncol = 3, byrow = TRUE)
m
TrimNA(m)
TrimNA(m, trim = "any")
}
