% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyFilter.R
\name{ApplyFilter}
\alias{ApplyFilter}
\title{Apply a filter to a timeseries}
\usage{
ApplyFilter(data, filter, method = 1)
}
\arguments{
\item{data}{Input timeseries (ts object)}

\item{filter}{vector of filter weights}

\item{method}{constraint method choice 1-3}
}
\value{
filtered timeseries (ts object)
}
\description{
Apply a filter to a timeseries
}
\details{
Using endpoint constrains as describen in  Mann et al., GRL 2003
minimum norm constraint (method=1)
minimum slope constraint (method=2)
minimum roughness constraint (method=3)
}
\author{
Thomas Laepple
}

