% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossSpecMTM.R
\name{CrossSpecMTM}
\alias{CrossSpecMTM}
\title{MTM cross spectral estimator}
\usage{
CrossSpecMTM(x, y = NULL, dt = 1, k = NULL, nw = 2, detrend = T)
}
\arguments{
\item{x, y}{input data vectors (or array if y=NULL)}

\item{dt}{time stepping (default = 1)}

\item{k}{The number of Slepian tapers used, often (default = NULL). For
default NULL (2*nw -1) is used except for the cases 2*nw > N}

\item{nw}{slepian bandwidth parameter, typically something between 2.0
(default) and 8.0}

\item{detrend}{(default=T)}
}
\value{
spectra object
}
\description{
estimates cross spectral densities using slepian tapers with
  adaptive weighting. Function is an adaptation of Peter Huybers Matlab
  coherence function cmtm.m
}
\examples{
x <- SimPowerlaw(1, 1e03)
y <- ApplyFilter(x, filter = dnorm(-5:5, 0, 1), method  =3)
cs1 <- CrossSpecMTM(x, y)
plot(cs1$freq[1:length(x)], cs1$coh, log = "xy", type = "l")

ch1 <- Coherence(x, y)
lines(ch1$freq, ch1$coh, col = "red")

}
