% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossSpecMTM.R
\name{CrossSpecMTM}
\alias{CrossSpecMTM}
\title{MTM cross spectral estimator}
\usage{
CrossSpecMTM(x, y = NULL, dt = 1, NW = 8, k = NULL, dpssIN = NULL, detrend = T)
}
\arguments{
\item{x, y}{input data vectors (or array if y=NULL)}

\item{dt}{time stepping (default = 1)}

\item{NW}{number of slepian taper (default = 8)}

\item{k}{order (default = NULL)}

\item{dpssIN}{allows to use an predefined dpss object to reduce computation time when applied multiple times (default=NULL). If specified NW and k are obsolete.}

\item{detrend}{(default=T)}
}
\value{
spectra object
}
\description{
estimates cross spectral densities using slepian tapers with adaptive weighting. Function is an adaption of Peter Huybers Matlab coherence function cmtm.m
}
