% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTransferFunction.R
\name{GetTransferFunction}
\alias{GetTransferFunction}
\title{Derives and plots the transfer function}
\usage{
GetTransferFunction(g.u, resolution = 100, bPlot = TRUE, add = FALSE, ...)
}
\arguments{
\item{g.u}{}

\item{resolution}{}

\item{bPlot}{}

\item{add}{}

\item{...}{}
}
\value{
list(omega,y) containing the transfer function
}
\description{
Derives and plots the transfer function (given a filter)
}
\details{
Get the transfer function of a symetric filter, page 122 in Bloomfield 1976,
}
\author{
Thomas Laepple
}

