% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPlot.R
\name{LLines}
\alias{LLines}
\title{Add a spectrum to an existing log-log spectral plot.}
\usage{
LLines(
  x,
  conf = TRUE,
  bPeriod = FALSE,
  col = "black",
  alpha = 0.3,
  removeFirst = 0,
  removeLast = 0,
  ...
)
}
\arguments{
\item{x}{a spectral object resulting from a call to \code{\link{SpecMTM}}.}

\item{conf}{if \code{TRUE} (the default) add a transparent confidence
interval (suppressed if \code{x} contains no error limits).}

\item{bPeriod}{if \code{TRUE} the x-axis is displayed in units of period
(inverse frequency), increasing to the left. Defaults to \code{FALSE}.}

\item{col}{color for the line plot and the confidence interval.}

\item{alpha}{transparency level (between 0 and 1) for the confidence
interval. Defaults to \code{0.3}.}

\item{removeFirst}{omit \code{removeFirst} values on the low frequency side.}

\item{removeLast}{omit \code{removeLast} values on the high frequency side.}

\item{...}{further graphical parameters passed to \code{lines}.}
}
\description{
This function adds a spectrum to an existing double-logarithmic plot and
optionally adds a transparent confidence interval.
}
\examples{
x <- ts(arima.sim(list(ar = 0.9), 1000))
spec <- SpecMTM(x)
LPlot(spec, col = "grey")
LLines(LogSmooth(spec), lwd = 2)
}
\seealso{
Other functions to plot power spectra: 
\code{\link{LPlot}()},
\code{\link{gg_spec}()}
}
\author{
Thomas Laepple
}
\concept{functions to plot power spectra}
