% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLines.R
\name{LLines}
\alias{LLines}
\title{add Logplot + transparent confidence interval for the spectral plotting}
\usage{
LLines(x, conf = TRUE, col = "black", alpha = 0.3, removeFirst = 0,
  removeLast = 0, ...)
}
\arguments{
\item{x}{spectra object}

\item{conf}{TRUE: Plot confidence interval}

\item{col}{color}

\item{alpha}{transparency}

\item{removeFirst}{omit removeFirst values on the low frequency side}

\item{removeLast}{omit removeFirst values on the high frequency side}

\item{...}{other parameters to be passed to the line functio}
}
\value{
none
}
\description{
add Logplot + transparent confidence interval for the spectral plotting
}
\author{
Thomas Laepple
}

