% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLines.R
\name{LLines}
\alias{LLines}
\title{add Logplot + transparent confidence interval for the spectral plotting}
\usage{
LLines(x, conf = TRUE, bPeriod = FALSE, col = "black", alpha = 0.3,
  removeFirst = 0, removeLast = 0, ...)
}
\arguments{
\item{x}{spectra object}

\item{conf}{TRUE: Plot confidence interval}

\item{bPeriod}{if \code{TRUE} treat the x-axis values in units of period
(inverse frequency). Defaults to \code{FALSE}.}

\item{col}{color}

\item{alpha}{transparency}

\item{removeFirst}{omit removeFirst values on the low frequency side}

\item{removeLast}{omit removeFirst values on the high frequency side}

\item{...}{other parameters to be passed to the line function}
}
\value{
none
}
\description{
add Logplot + transparent confidence interval for the spectral plotting
}
\examples{
x<-ts(arima.sim(list(ar = 0.9),1000))
spec<-SpecMTM(x)
LPlot(spec,col="grey")
LLines(LogSmooth(spec),lwd=3)
}
\author{
Thomas Laepple
}
