% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecInterpolate.R
\name{SpecInterpolate}
\alias{SpecInterpolate}
\title{Interpolates the spectrum spec to the specRef frequency resolution}
\usage{
SpecInterpolate(freqRef, spec)
}
\arguments{
\item{freqRef}{frequency vector of the target resolution}

\item{spec}{list(spec,freq,dof)}
}
\value{
one spectum as list(spec,freq,dof) (spec on the specRef resolution)
}
\description{
Interpolates the spectrum spec to the specRef frequency resolution
}
\author{
Thomas Laepple
}
