% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecInterpolate.R
\name{SpecInterpolate}
\alias{SpecInterpolate}
\title{Interpolate spectrum}
\usage{
SpecInterpolate(freqRef, spec)
}
\arguments{
\item{freqRef}{numeric vector with the target frequency resolution.}

\item{spec}{a spectral object of class \code{"spec"} or a list with the
minimum components \code{freq}, \code{spec} and \code{dof} which are
vectors of the same length giving the original frequency resolution and the
corresponding spectral estimates and degrees of freedom.}
}
\value{
a spectral object of class \code{"spec"} with the spectrum
  interpolated to the target resolution.
}
\description{
Interpolate a spectrum to a given frequency resolution. The interpolation
includes the spectral estimates themselves as well as their degrees of
freedom.
}
\examples{

freqRef <- seq(0.1, 0.5, 0.1)
spec <- list(freq = c(0.1, 0.2, 0.4, 0.5), spec = c(1, 2, 4, 5),
             dof = rep(1, 4))

SpecInterpolate(freqRef, spec)
}
\author{
Thomas Laepple
}
