% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecMTM.R
\name{SpecMTM}
\alias{SpecMTM}
\title{MTM spectral estimator}
\usage{
SpecMTM(timeSeries, k = 3, nw = 2, nFFT = "default",
  centre = c("Slepian"), dpssIN = NULL, returnZeroFreq = FALSE,
  Ftest = FALSE, jackknife = FALSE, jkCIProb = 0.95,
  maxAdaptiveIterations = 100, plot = FALSE, na.action = na.fail,
  returnInternals = FALSE, detrend = TRUE, bPad = FALSE, ...)
}
\arguments{
\item{timeSeries}{A time series of equally spaced data, this can be created
by the ts() function where deltat is specified.}

\item{k}{a positive integer, the number of tapers, often 2*nw.}

\item{nw}{a positive double precision number, the time-bandwidth
parameter.}

\item{nFFT}{This function pads the data before computing the fft. nFFT
indicates the total length of the data after padding.}
}
\value{
spectra object list(freq, spec, dof)
}
\description{
MTM spectral estimator
calls \code{\link[multitaper]{spec.mtm}} from library multitaper
}
\examples{
x <- ts(arima.sim(list(ar = 0.9), 1000))
spec <- SpecMTM(x)
LPlot(spec, col='grey')
LLines(LogSmooth(spec), lwd=2)
}
\author{
Thomas Laepple
}

