% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecMTM.R
\name{SpecMTM}
\alias{SpecMTM}
\title{MTM spectral estimator}
\usage{
SpecMTM(timeSeries, k = 3, nw = 2, nFFT = "default",
  centre = c("Slepian"), dpssIN = NULL, returnZeroFreq = FALSE,
  Ftest = FALSE, jackknife = FALSE, jkCIProb = 0.95,
  maxAdaptiveIterations = 100, plot = FALSE, na.action = na.fail,
  returnInternals = FALSE, detrend = TRUE, bPad = FALSE, ...)
}
\arguments{
\item{timeSeries}{A time series of equally spaced data, this can be created
by the ts() function where deltat is specified.}

\item{k}{a positive integer, the number of tapers, often 2*nw.}

\item{nw}{a positive double precision number, the time-bandwidth
parameter.}

\item{nFFT}{This function pads the data before computing the fft. nFFT
indicates the total length of the data after padding.}

\item{centre}{}

\item{dpssIN}{}

\item{returnZeroFreq}{}

\item{Ftest}{}

\item{jackknife}{}

\item{jkCIProb}{}

\item{maxAdaptiveIterations}{}

\item{plot}{}

\item{na.action}{}

\item{returnInternals}{}

\item{detrend}{}

\item{bPad}{}

\item{...}{}
}
\value{
spectra object list(freq,spec,dof)
}
\description{
MTM spectral estimator
calls spec.mtm from library multitaper
 see spec.mtm from library mulitaper ?spec.mtm
}
\author{
Thomas Laepple
}

