% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecACF.R
\name{TrimNA}
\alias{TrimNA}
\title{Remove leading and trailing rows of all NA}
\usage{
TrimNA(m)
}
\arguments{
\item{m}{a numeric matrix}
}
\value{
a numeric matrix
}
\description{
Remove leading and trailing rows of all NA
}
\examples{
m <- matrix(c(NA, NA, NA, 1:9, NA,NA,NA, 10:12, NA,NA,NA), ncol = 3, byrow = TRUE)
m
PaleoSpec:::TrimNA(m)
}
\keyword{internal}
